backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
spacy==3.6.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
requests==2.31.0
torch
torchtext
joblib>=1.0.1
python-dotenv==0.19.0
redis>=3.5.3
transformers>=4.30.0
google-generativeai>=0.3.0
tenacity==8.0.1